refactor: Move `IFluidErrorBase` to `core-interfaces`
